feat(ui): organise gallery slice
