+++ invokeai/frontend/web/src/features/gallery/store/gallerySlice.ts
@@ -1,19 +1,15 @@
 import type { PayloadAction, Update } from '@reduxjs/toolkit';
-import {
-  createEntityAdapter,
-  createSelector,
-  createSlice,
-} from '@reduxjs/toolkit';
+import { createEntityAdapter, createSlice } from '@reduxjs/toolkit';
 import { RootState } from 'app/store/store';
-import { defaultSelectorOptions } from 'app/store/util/defaultMemoizeOptions';
 import { dateComparator } from 'common/util/dateComparator';
-import { keyBy, uniq } from 'lodash-es';
+import { uniq } from 'lodash-es';
 import { boardsApi } from 'services/api/endpoints/boards';
 import {
   imageUrlsReceived,
   receivedPageOfImages,
 } from 'services/api/thunks/image';
 import { ImageCategory, ImageDTO } from 'services/api/types';
+import { selectFilteredImagesLocal } from './gallerySelectors';
 
 export const imagesAdapter = createEntityAdapter<ImageDTO>({
   selectId: (image) => image.image_name,
@@ -224,45 +220,3 @@ export const {
 } = gallerySlice.actions;
 
 export default gallerySlice.reducer;
-
-export const selectFilteredImagesLocal = createSelector(
-  (state: typeof initialGalleryState) => state,
-  (galleryState) => {
-    const allImages = imagesAdapter.getSelectors().selectAll(galleryState);
-    const { categories, selectedBoardId } = galleryState;
-
-    const filteredImages = allImages.filter((i) => {
-      const isInCategory = categories.includes(i.image_category);
-      const isInSelectedBoard = selectedBoardId
-        ? i.board_id === selectedBoardId
-        : true;
-      return isInCategory && isInSelectedBoard;
-    });
-
-    return filteredImages;
-  }
-);
-
-export const selectFilteredImages = createSelector(
-  (state: RootState) => state,
-  (state) => {
-    return selectFilteredImagesLocal(state.gallery);
-  },
-  defaultSelectorOptions
-);
-
-export const selectFilteredImagesAsObject = createSelector(
-  selectFilteredImages,
-  (filteredImages) => keyBy(filteredImages, 'image_name')
-);
-
-export const selectFilteredImagesIds = createSelector(
-  selectFilteredImages,
-  (filteredImages) => filteredImages.map((i) => i.image_name)
-);
-
-export const selectLastSelectedImage = createSelector(
-  (state: RootState) => state,
-  (state) => state.gallery.selection[state.gallery.selection.length - 1],
-  defaultSelectorOptions
-);
